Delete HTTPTransport, inline health checks into GatewayConnectionManager
HTTPTransport was a 490-line intermediary that only did health checks
and hosted bash/file execution extensions. Both responsibilities now
have better homes:

- Health checks (~100 lines): inlined into GatewayConnectionManager,
  which already owns the connection lifecycle. No more separate transport
  object with callback wiring.

- Host bash/file execution (~335 lines): moved to standalone
  HostToolExecutor enum. These are self-contained operations that spawn
  processes and post results via HostProxyClient — they never needed
  transport state.

- Response type structs (AssistantEvent, ConversationsListResponse,
  workspace types): moved to NetworkResponseTypes.swift since they're
  used across multiple clients.

Consumers that accessed httpTransport.baseURL/bearerToken now read
from config.transport directly. AppDelegate subscribe loop calls
HostToolExecutor instead of httpTransport.

Files deleted:
- HTTPDaemonClient.swift (490 lines)
- HTTPDaemonClient+HostBash.swift (159 lines)
- HTTPDaemonClient+HostFile.swift (176 lines)

Files created:
- HostToolExecutor.swift (305 lines)
- NetworkResponseTypes.swift (85 lines)

Net: -162 lines
