+++ clients/shared/Network/GatewayConnectionManager.swift
@@ -3,21 +3,23 @@ import os
 
 private let log = Logger(subsystem: Bundle.main.bundleIdentifier ?? "com.vellum.vellum-assistant", category: "GatewayConnectionManager")
 
+/// Minimal decode of the healthz response to extract the version field.
+private struct HealthzVersionResponse: Decodable {
+    let version: String?
+}
+
 /// Manages the gateway connection lifecycle: connect, disconnect,
-/// reconfigure, auto-wake, and bearer token updates.
+/// reconfigure, auto-wake, health checks, and bearer token updates.
 ///
-/// Owns `HTTPTransport` (health checks) and coordinates with `EventStreamClient`
-/// (SSE start/stop). Does NOT own published state — `DaemonStatus` observes
-/// this manager's callbacks to update its `@Published` properties.
+/// Coordinates with `EventStreamClient` (SSE start/stop). Does NOT own
+/// published state — `DaemonStatus` observes this manager's callbacks
+/// to update its `@Published` properties.
 @MainActor
 public final class GatewayConnectionManager {
 
-    // MARK: - Transport
+    // MARK: - Connection State
 
-    /// HTTP transport for health checks and host tool execution.
-    public var httpTransport: HTTPTransport?
-
-    /// Current daemon connection configuration.
+    /// Current connection configuration.
     public private(set) var config: DaemonConfig
 
     /// Whether a connection attempt is in flight.
@@ -25,6 +27,58 @@ public final class GatewayConnectionManager {
 
     /// Whether the transport has authenticated successfully.
     var isAuthenticated = false
+
+    /// Base URL of the connected gateway/runtime.
+    private var baseURL: String?
+
+    /// Bearer token for authentication.
+    private var bearerToken: String?
+
+    /// Transport metadata (route mode, auth mode, platform assistant ID).
+    private var transportMetadata: TransportMetadata?
+
+    // MARK: - Health Check
+
+    /// Periodic health check task.
+    private var healthCheckTask: Task<Void, Never>?
+
+    /// Health check interval in seconds.
+    private let healthCheckInterval: TimeInterval = 15.0
+
+    /// Whether the gateway is reachable (health check passes).
+    private(set) var isConnected: Bool = false
+
+    /// The daemon's self-reported version from the most recent health check.
+    private(set) var daemonVersion: String?
+
+    /// Whether we should attempt to reconnect on disconnect.
+    private var shouldReconnect = true
+
+    /// Whether a planned service group update is in progress.
+    /// Accelerates health check polling for faster reconnection.
+    var isUpdateInProgress: Bool = false
+
+    /// Update the in-progress flag and restart the health-check loop when
+    /// transitioning to `true`.
+    func setUpdateInProgress(_ value: Bool) {
+        let wasInProgress = isUpdateInProgress
+        isUpdateInProgress = value
+        if value && !wasInProgress && healthCheckTask != nil {
+            startHealthCheckLoop()
+        }
+    }
+
+    // MARK: - 401 Recovery
+
+    /// Result of an async authentication refresh attempt.
+    enum AuthRefreshResult {
+        case success
+        case transientFailure
+        case terminalFailure
+    }
+
+    /// In-flight refresh task for coalescing concurrent 401 handlers.
+    private var refreshTask: Task<AuthRefreshResult, Never>?
 
     // MARK: - Auto-Wake
 
@@ -39,9 +93,7 @@ public final class GatewayConnectionManager {
     public var recoveryDeviceId: String?
 
     #if os(macOS)
-    /// Timestamp of the last auto-wake attempt.
     var lastAutoWakeAttempt: Date?
-    /// The in-flight auto-wake task.
     var autoWakeTask: Task<Void, Never>?
     #endif
 
@@ -58,7 +110,7 @@ public final class GatewayConnectionManager {
 
     // MARK: - Dependencies
 
-    /// Event stream client — ConnectionManager starts/stops SSE and registers conversation IDs.
+    /// Event stream client — starts/stops SSE and registers conversation IDs.
     private let eventStreamClient: EventStreamClient
 
     // MARK: - Init
@@ -85,49 +137,29 @@ public final class GatewayConnectionManager {
             return
         }
 
-        log.info("connect: establishing HTTP transport to \(baseURL, privacy: .public)")
+        log.info("connect: establishing connection to \(baseURL, privacy: .public)")
 
-        let transport = HTTPTransport(
-            baseURL: baseURL,
-            bearerToken: bearerToken,
-            conversationKey: conversationKey,
-            transportMetadata: config.transportMetadata
-        )
-
-        // Bridge HTTP transport connection state to DaemonStatus via callback.
-        transport.onConnectionStateChanged = { [weak self] connected in
-            guard let self else { return }
-            self.isConnecting = false
-            self.onConnectionStateChanged?(connected)
-            #if os(macOS)
-            if !connected {
-                self.autoWakeIfDaemonDied()
-            }
-            #endif
-        }
-
-        // Sync daemon version from health checks.
-        transport.onDaemonVersionChanged = { [weak self] newVersion in
-            self?.onDaemonVersionChanged?(newVersion)
-        }
-
-        // Broadcast auth errors from health check 401 handling.
-        transport.onAuthError = { [weak self] message in
-            self?.onAuthError?(message)
-        }
-
-        self.httpTransport = transport
+        self.baseURL = baseURL.hasSuffix("/") ? String(baseURL.dropLast()) : baseURL
+        self.bearerToken = bearerToken
+        self.transportMetadata = config.transportMetadata
 
         // Register the conversation key for host tool filtering
         if !conversationKey.isEmpty {
             eventStreamClient.registerConversationId(conversationKey)
         }
 
+        shouldReconnect = true
+
         do {
-            try await transport.connect()
+            // Run initial health check
+            try await performHealthCheck()
+
+            // Start periodic health checks
+            startHealthCheckLoop()
+
             isAuthenticated = true
             isConnecting = false
-            log.info("connect: transport connected successfully to \(baseURL, privacy: .public)")
+            log.info("connect: connected successfully to \(baseURL, privacy: .public)")
 
             // Auto-start SSE now that health check passed
             eventStreamClient.startSSE()
@@ -135,7 +167,7 @@ public final class GatewayConnectionManager {
             #if os(macOS)
             guard !Task.isCancelled else {
                 isConnecting = false
-                httpTransport = nil
+                self.baseURL = nil
                 log.info("connect: task cancelled — skipping auto-wake")
                 throw error
             }
@@ -147,7 +179,8 @@ public final class GatewayConnectionManager {
                     do {
                         try await wakeHandler()
                         log.info("connect: auto-wake succeeded, retrying connection to \(baseURL, privacy: .public)")
-                        try await transport.connect()
+                        try await performHealthCheck()
+                        startHealthCheckLoop()
                         isAuthenticated = true
                         isConnecting = false
                         log.info("connect: retry after auto-wake succeeded for \(baseURL, privacy: .public)")
@@ -161,8 +194,8 @@ public final class GatewayConnectionManager {
             #endif
 
             isConnecting = false
-            httpTransport = nil
-            log.error("connect: transport connection failed for \(baseURL, privacy: .public): \(error)")
+            self.baseURL = nil
+            log.error("connect: connection failed for \(baseURL, privacy: .public): \(error)")
             throw error
         }
     }
@@ -183,18 +216,17 @@ public final class GatewayConnectionManager {
         }
         #endif
 
-        httpTransport?.disconnect()
-        httpTransport = nil
-        // Stop SSE but preserve subscriber streams — consumers start one-shot
-        // for-await loops that must survive reconnects and assistant switches.
+        shouldReconnect = false
+        healthCheckTask?.cancel()
+        healthCheckTask = nil
+        setConnected(false)
+
+        // Stop SSE but preserve subscriber streams
         eventStreamClient.stopSSE()
     }
 
     // MARK: - Reconfigure
 
-    /// Reconfigure the transport in place without replacing object identity.
-    /// Preserves subscriber references across assistant switches.
-    /// Callers must call `connect()` after reconfiguring.
     public func reconfigure(config newConfig: DaemonConfig) {
         #if os(macOS)
         autoWakeTask?.cancel()
@@ -203,6 +235,9 @@ public final class GatewayConnectionManager {
         disconnect()
         self.config = newConfig
         isAuthenticated = false
+        baseURL = nil
+        bearerToken = nil
+        transportMetadata = nil
         #if os(macOS)
         lastAutoWakeAttempt = nil
         #endif
@@ -210,9 +245,200 @@ public final class GatewayConnectionManager {
 
     // MARK: - Token Update
 
-    /// Push a new bearer token to the active HTTP transport.
     public func updateBearerToken(_ token: String) {
-        httpTransport?.updateBearerToken(token)
+        bearerToken = token
+    }
+
+    // MARK: - Health Check
+
+    /// Build the health check URL based on the current route mode.
+    private func buildHealthCheckURL() -> URL? {
+        guard let baseURL else { return nil }
+
+        let path: String
+        switch transportMetadata?.routeMode {
+        case .platformAssistantProxy:
+            guard let assistantId = transportMetadata?.platformAssistantId else {
+                log.error("platformAssistantProxy route mode requires platformAssistantId")
+                return nil
+            }
+            path = "/v1/assistants/\(assistantId)/healthz/"
+        default:
+            path = "/healthz"
+        }
+
+        return URL(string: "\(baseURL)\(path)")
+    }
+
+    /// Run a single health check against the gateway.
+    private func performHealthCheck() async throws {
+        guard let healthURL = buildHealthCheckURL() else {
+            throw ConnectionError.invalidURL
+        }
+        var healthReq = URLRequest(url: healthURL)
+        healthReq.timeoutInterval = 10
+        applyAuth(&healthReq)
+
+        do {
+            let (data, response) = try await URLSession.shared.data(for: healthReq)
+            guard let http = response as? HTTPURLResponse, http.statusCode == 200 else {
+                let statusCode = (response as? HTTPURLResponse)?.statusCode ?? -1
+                if statusCode == 401 {
+                    handleAuthenticationFailure(responseData: data)
+                    if isManagedMode {
+                        shouldReconnect = false
+                    }
+                }
+                throw ConnectionError.healthCheckFailed
+            }
+            // Extract daemon version from response body
+            if let decoded = try? JSONDecoder().decode(HealthzVersionResponse.self, from: data) {
+                if let newVersion = decoded.version, newVersion != daemonVersion {
+                    daemonVersion = newVersion
+                    if let id = UserDefaults.standard.string(forKey: "connectedAssistantId"), !id.isEmpty {
+                        LockfilePaths.updateServiceGroupVersion(assistantId: id, version: newVersion)
+                    }
+                    onDaemonVersionChanged?(newVersion)
+                } else if let newVersion = decoded.version {
+                    daemonVersion = newVersion
+                }
+            }
+            log.info("Health check passed for \(self.baseURL ?? "?", privacy: .public)")
+            setConnected(true)
+        } catch let error as ConnectionError {
+            setConnected(false)
+            throw error
+        } catch {
+            log.error("Health check failed: \(error.localizedDescription)")
+            setConnected(false)
+            throw ConnectionError.healthCheckFailed
+        }
+    }
+
+    /// Periodically poll `/healthz` to maintain connection status.
+    private func startHealthCheckLoop() {
+        healthCheckTask?.cancel()
+
+        healthCheckTask = Task { @MainActor [weak self] in
+            while !Task.isCancelled {
+                do {
+                    let interval = (self?.isUpdateInProgress == true) ? 2.0 : (self?.healthCheckInterval ?? 15.0)
+                    try await Task.sleep(nanoseconds: UInt64(interval * 1_000_000_000))
+                } catch {
+                    return
+                }
+
+                guard let self, self.shouldReconnect else { return }
+
+                do {
+                    try await self.performHealthCheck()
+                } catch {
+                    log.warning("Periodic health check failed: \(error.localizedDescription)")
+                }
+            }
+        }
+    }
+
+    // MARK: - 401 Recovery
+
+    private func handleAuthenticationFailure(responseData: Data? = nil) {
+        if isManagedMode {
+            log.warning("401 in managed mode — session token may be expired")
+            onAuthError?(.conversationError(ConversationErrorMessage(
+                conversationId: "",
+                code: .authenticationRequired,
+                userMessage: "Session expired. Please sign in again.",
+                retryable: false
+            )))
+            disconnect()
+            return
+        }
+
+        Task { @MainActor [weak self] in
+            guard let self else { return }
+            _ = await self.handleAuthenticationFailureAsync(responseData: responseData)
+        }
+    }
+
+    func handleAuthenticationFailureAsync(responseData: Data? = nil) async -> AuthRefreshResult {
+        if isManagedMode {
+            log.warning("401 in managed mode — session token may be expired")
+            onAuthError?(.conversationError(ConversationErrorMessage(
+                conversationId: "",
+                code: .authenticationRequired,
+                userMessage: "Session expired. Please sign in again.",
+                retryable: false
+            )))
+            disconnect()
+            return .terminalFailure
+        }
+
+        let terminalCodes: Set<String> = ["credentials_revoked"]
+        if let data = responseData,
+           let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any] {
+            let code = (json["error"] as? [String: Any])?["code"] as? String
+            if let code, terminalCodes.contains(code) {
+                log.error("Terminal 401 code: \(code) — re-auth required")
+                self.onAuthError?(.conversationError(ConversationErrorMessage(
+                    conversationId: "",
+                    code: .authenticationRequired,
+                    userMessage: "Session expired. Please re-pair your device.",
+                    retryable: false
+                )))
+                return .terminalFailure
+            }
+        }
+
+        if let existing = refreshTask {
+            return await existing.value
+        }
+
+        let task = Task<AuthRefreshResult, Never> { @MainActor [weak self] in
+            guard let self else { return .transientFailure }
+            defer { self.refreshTask = nil }
+            return await self.performRefresh()
+        }
+        refreshTask = task
+        return await task.value
+    }
+
+    private func performRefresh() async -> AuthRefreshResult {
+        guard let baseURL else { return .transientFailure }
+
+        #if os(macOS)
+        let refreshPlatform = "macos"
+        let refreshDeviceId = HostIdComputer.computeHostId()
+        #else
+        let refreshPlatform = "ios"
+        let refreshDeviceId = APIKeyManager.shared.getAPIKey(provider: "pairing-device-id") ?? ""
+        #endif
+
+        let result = await ActorCredentialRefresher.refresh(
+            baseURL: baseURL,
+            bearerToken: self.bearerToken,
+            platform: refreshPlatform,
+            deviceId: refreshDeviceId
+        )
+
+        switch result {
+        case .success:
+            log.info("Token refresh succeeded")
+            return .success
+
+        case .terminalError(let reason):
+            log.error("Token refresh failed terminally: \(reason) — re-pair required")
+            self.onAuthError?(.conversationError(ConversationErrorMessage(
+                conversationId: "",
+                code: .authenticationRequired,
+                userMessage: "Session expired. Please re-pair your device.",
+                retryable: false
+            )))
+            return .terminalFailure
+
+        case .transientError:
+            log.warning("Token refresh encountered transient error — will retry on next 401")
+            return .transientFailure
+        }
     }
 
     // MARK: - Auto-Wake
@@ -263,6 +489,59 @@ public final class GatewayConnectionManager {
         }
     }
     #endif
+
+    // MARK: - Helpers
+
+    private func applyAuth(_ request: inout URLRequest) {
+        guard let transportMetadata else { return }
+        switch transportMetadata.authMode {
+        case .bearerToken:
+            if let accessToken = ActorTokenManager.getToken(), !accessToken.isEmpty {
+                request.setValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
+            } else if let token = bearerToken {
+                request.setValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
+            }
+        case .sessionToken:
+            if let token = SessionTokenManager.getToken() {
+                request.setValue(token, forHTTPHeaderField: "X-Session-Token")
+            }
+            if let orgId = UserDefaults.standard.string(forKey: "connectedOrganizationId") {
+                request.setValue(orgId, forHTTPHeaderField: "Vellum-Organization-Id")
+            }
+        }
+    }
+
+    private var isManagedMode: Bool {
+        transportMetadata?.routeMode == .platformAssistantProxy
+    }
+
+    private func setConnected(_ connected: Bool) {
+        guard isConnected != connected else { return }
+        isConnected = connected
+        isConnecting = false
+        onConnectionStateChanged?(connected)
+        #if os(macOS)
+        if !connected {
+            autoWakeIfDaemonDied()
+        }
+        #endif
+    }
+
+    // MARK: - Errors
+
+    public enum ConnectionError: Error, LocalizedError {
+        case healthCheckFailed
+        case invalidURL
+
+        public var errorDescription: String? {
+            switch self {
+            case .healthCheckFailed:
+                return "Gateway health check failed"
+            case .invalidURL:
+                return "Invalid gateway URL"
+            }
+        }
+    }
 
     deinit {
         #if os(macOS)
